refactor: rename namespace to `errors`
